move `Tiles` to the top level of `layout`
