+++ src/core/components/layout/Tiles/styles.ts
@@ -1,5 +1,5 @@
 import { css, SerializedStyles } from '@emotion/react';
-import { Columns } from './tiles';
+import { Columns } from './Tiles';
 import { from, until } from '@guardian/src-foundations/mq';
 import { space } from '@guardian/src-foundations';
