+++ src/core/components/layout/index.tsx
@@ -2,5 +2,5 @@ export { Columns, Column } from './components/columns/columns';
 export { Container, ContainerProps } from './Container/Container';
 export { Hide, HideProps } from './Hide/Hide';
 export { Stack } from './Stack/Stack';
-export { Tiles } from './components/tiles/tiles';
-export { Inline } from './Inline/Inline';
+export { Tiles, TilesProps } from './Tiles/Tiles';
+export { Inline, InlineProps } from './Inline/Inline';
